+++ src/main/java/com/magento/idea/magento2plugin/actions/generation/generator/OverrideTemplateInThemeGenerator.java
@@ -70,7 +70,7 @@ public class OverrideTemplateInThemeGenerator extends OverrideInThemeGenerator {
         }
 
         final ModuleIndex moduleIndex = new ModuleIndex(project);
-        PsiDirectory directory = moduleIndex.getModuleDirectoryByModuleName(themeName);
+        PsiDirectory directory = moduleIndex.getThemeDirectoryByThemeName(themeName);
 
         if (directory == null) {
             return;